pyproject.toml
[tool.poetry]
name = "qa-guru-hw19-browserstack"
version = "0.1.0"
description = ""
authors = ["Zaur Mamedov <zaurmd1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selene = "^2.0.0rc8"
pytest = "^8.1.1"
allure-pytest = "^2.13.5"
appium-python-client = "^4.0.0"
python-dotenv = "^0.20.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
